Refactor code base
