+++ friends/src/redux/actions/types/index.js
@@ -1,0 +1,6 @@
+// Action Types
+export const SET_ACTION_STATUS = 'SET_ACTION_STATUS';
+export const DELETE_FRIEND = 'DELETE_FRIEND';
+export const ADD_FRIENDS = 'ADD_FRIENDS';
+export const ADD_FRIEND = 'ADD_FRIEND';
+export const UPDATE_FRIEND_INFO = 'UPDATE_FRIEND_INFO';
